+++ firmware/hexray/BMS/src/app/segments/app_segments_reach.cpp
@@ -8,19 +8,7 @@ using io::adbms::Segments;
 
 Segments<result<bool>> check()
 {
-    const io::adbms::SpiBusReach bus = io::adbms::misc::getSpiBusReach();
-
-    // A segment is reachable if it falls within reach of either end of the
-    // daisy chain: from the low side (segments [0, ls_reach)) or from the high
-    // side (segments [NUM_SEGMENTS - hs_reach, NUM_SEGMENTS)).
-    Segments<result<bool>> per_seg;
-    for (size_t seg = 0U; seg < NUM_SEGMENTS; seg++)
-    {
-        const bool reachable_from_ls = seg < bus.ls_reach;
-        const bool reachable_from_hs = seg >= static_cast<size_t>(NUM_SEGMENTS - bus.hs_reach);
-        per_seg[seg]                 = reachable_from_ls || reachable_from_hs;
-    }
-    return per_seg;
+    return {};
 }
 
 } // namespace app::segments::reach
