--- firmware/hexray/BMS/src/io/adbms/io_adbms_protocol.cpp
@@ -1,11 +1,0 @@
-#include "io_adbms_protocol.hpp"
-
-namespace io::adbms
-{
-SemaphoreHandle_t spiMutex()
-{
-    static StaticSemaphore_t storage;
-    static SemaphoreHandle_t handle = xSemaphoreCreateRecursiveMutexStatic(&storage);
-    return handle;
-}
-} // namespace io::adbms
